changed readings to floats for better accuracy

+++ c_func.h
@@ -2,6 +2,6 @@
 void input_setup(void);
 int read_input(void);
 void ada1893_setup(int alt_pres);
-int read_temp();
-int read_alt();
-int read_pres();
+float read_temp();
+float read_alt();
+float read_pres();
